refactor: move universal git utils to common-all
