+++ packages/common-server/src/git.ts
@@ -41,14 +41,7 @@ export class GitUtils {
     remotePath: string;
     accessToken: string;
   }) {
-    const { remotePath, accessToken } = opts;
-    let repoPath: string;
-    if (remotePath.startsWith("https://")) {
-      repoPath = remotePath.split("/").slice(-2).join("/");
-    } else {
-      repoPath = opts.remotePath.split(":").slice(-1)[0];
-    }
-    return `https://${accessToken}:x-oauth-basic@github.com/${repoPath}`;
+    return CommonGitUtils.getGithubAccessTokenUrl(opts);
   }
 
   static git2Github(gitUrl: string) {
@@ -64,12 +57,11 @@ export class GitUtils {
   }
 
   static getOwnerAndRepoFromURL(url: string) {
-    const [owner, repo] = url.split("/").slice(-2);
-    return { owner, repo };
+    return CommonGitUtils.getOwnerAndRepoFromURL(url);
   }
 
-  static getRepoNameFromURL(url: string): string {
-    return path.basename(url, ".git");
+  static getRepoNameFromURL(url: string) {
+    return CommonGitUtils.getRepoNameFromURL(url);
   }
 
   /** Find the dependency path for a vault given the remote url. You can use
@@ -86,57 +78,7 @@ export class GitUtils {
     vaultName: string;
     url?: string;
   }): string {
-    // If no remote URL exists, then it's a local vault. We keep these in a
-    // local-only folder.
-    if (url === undefined) return FOLDERS.LOCAL_DEPENDENCY;
-    // Check if it matches any web URLs like
-    // https://github.com/dendronhq/dendron-site.git This may also look like
-    // http://example.com:8000/dendronhq/dendron-site.git, we skip the port
-    const webMatch =
-      // starts with http:// or https://
-      // followed by the domain, which will continue until we hit /
-      // if we see a port definition like :8000, we skip it for simplicity
-      // then we have the path of the URL, like dendronhq/dendron-site
-      // finally, if there's a `.git` we'll discard that for a cleaner name
-      /^(https?:\/\/)(?<domain>[^/:]+)(:[0-9]+)?\/(?<path>.+?)(\.git)?$/.exec(
-        url
-      );
-    if (webMatch?.groups?.domain && webMatch?.groups?.path) {
-      // matched a HTTP/S git remote
-      return path.join(
-        webMatch.groups.domain,
-        // Normalize for Windows so forward slashes are converted to backward ones
-        path.normalize(webMatch.groups.path)
-      );
-    }
-    // Check if it matches any SSH URLs like
-    // git@github.com:dendronhq/dendron-site.git This may also look like
-    // git@example.com:220/dendronhq/dendron-site.git, we skip the port and the
-    const sshMatch =
-      // SSH urls start with a user, like git@ or gitlab@, which we skip
-      // followed by the domain, which will continue until we hit :
-      // if we see a port definition like :8000, we skip it for simplicity
-      // then we have the path of the URL, like dendronhq/dendron-site
-      // this path may optionally begin with a /, which we'll skip
-      // finally, if there's a `.git` we'll discard that for a cleaner name
-      /^([^@]+@)(?<domain>[^:/]+):([0-9]+\/)?\/?(?<path>.+?)(\.git)?$/.exec(
-        url
-      );
-    if (sshMatch?.groups?.domain && sshMatch?.groups?.path) {
-      // matched a HTTP/S git remote
-      return path.join(
-        sshMatch.groups.domain,
-        // Normalize for Windows so forward slashes are converted to backward ones
-        path.normalize(sshMatch.groups.path)
-      );
-    }
-    // If none of these worked, try to make a fallback path. This may be because
-    // the remote points to a local directory, or because it's something we
-    // didn't expect.
-    const fallback = _.findLast(url.split(/[/\\]/), (part) => part.length > 0);
-    if (fallback) return fallback;
-    // Fallback for the fallback: if all else fails, just use the vault name
-    return vaultName;
+    return CommonGitUtils.remoteUrlToDependencyPath({ vaultName, url });
   }
 
   /** If this vault had this remote, what path should it be stored under?
@@ -150,29 +92,7 @@ export class GitUtils {
     remote: string | null;
     vault: DVault;
   }): string {
-    const vaultName =
-      vault.name ??
-      // if the vault has no name, compute one based on the path
-      _.findLast(vault.fsPath.split(/[/\\]/), (part) => part.length > 0) ??
-      // Fall back to fsPath directly if the calculation fails
-      vault.fsPath;
-
-    if (!remote) {
-      // local
-      return path.join(
-        FOLDERS.DEPENDENCIES,
-        FOLDERS.LOCAL_DEPENDENCY,
-        vaultName
-      );
-    } else {
-      return path.join(
-        FOLDERS.DEPENDENCIES,
-        GitUtils.remoteUrlToDependencyPath({
-          vaultName,
-          url: remote,
-        })
-      );
-    }
+    return CommonGitUtils.getDependencyPathWithRemote({ vault, remote });
   }
 
   static getVaultFromRepo(opts: {
@@ -180,11 +100,7 @@ export class GitUtils {
     repoUrl: string;
     wsRoot: string;
   }): DVault {
-    const { repoPath, wsRoot } = opts;
-    return {
-      fsPath: path.relative(wsRoot, repoPath),
-      remote: { type: "git", url: opts.repoUrl },
-    };
+    return CommonGitUtils.getVaultFromRepo(opts);
   }
 
   static async getVaultsFromRepo(opts: {
